[#20] Restructure project
