+++ servant-hmac-auth.cabal
@@ -23,6 +23,9 @@ source-repository head
 library
   hs-source-dirs:      src
   exposed-modules:     Servant.Auth.Hmac
+                           Servant.Auth.Hmac.Crypto
+                           Servant.Auth.Hmac.Client
+                           Servant.Auth.Hmac.Server
 
   build-depends:       base >= 4.11 && < 5
                      , base64-bytestring ^>= 1.0
