+++ src/Servant/Auth/Hmac.hs
@@ -1,356 +1,11 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE DataKinds           #-}
-{-# LANGUAGE FlexibleContexts    #-}
-{-# LANGUAGE InstanceSigs        #-}
-{-# LANGUAGE TypeFamilies        #-}
-{-# LANGUAGE ViewPatterns        #-}
-
--- | Signing functions.
+{- | Servant authentication with HMAC. Contains server and client
+implementation.
+-}
 
 module Servant.Auth.Hmac
-       ( -- * Data types
-         SecretKey (..)
-       , Signature (..)
-
-         -- * Signing
-         -- ** Raw signing
-       , sign
-       , signSHA256
-
-         -- ** Request signing
-       , RequestPayload (..)
-       , waiRequestToPayload
-       , servantRequestToPayload
-       , requestSignature
-       , verifySignatureHmac
-
-         -- * Servant
-         -- ** server
-       , HmacAuth
-       , HmacAuthContextHandlers
-       , HmacAuthContext
-       , hmacAuthServerContext
-       , hmacAuthHandler
-
-         -- ** client
-       , HmacClientM (..)
-       , runHmacClient
-       , hmacClient
-       , signRequestHmac
+       ( module Hmac
        ) where
 
-import Control.Monad ((>=>))
-import Control.Monad.Except (throwError)
-import Control.Monad.IO.Class (liftIO)
-import Control.Monad.Reader (MonadReader (..), ReaderT, asks, runReaderT)
-import Control.Monad.Trans.Class (lift)
-import Crypto.Hash (hash)
-import Crypto.Hash.Algorithms (MD5, SHA256)
-import Crypto.Hash.IO (HashAlgorithm)
-import Crypto.MAC.HMAC (HMAC (hmacGetDigest), hmac)
-import Data.Binary.Builder (toLazyByteString)
-import Data.ByteString (ByteString)
-import Data.CaseInsensitive (foldedCase)
-import Data.List (sort, uncons)
-import Data.Maybe (fromMaybe)
-import Data.Proxy (Proxy (..))
-import Data.Sequence ((<|))
-import Network.HTTP.Client (RequestBody (..))
-import Network.HTTP.Types (Header, HeaderName, Method, RequestHeaders)
-import Network.Wai (rawPathInfo, rawQueryString, requestBody, requestHeaderHost, requestMethod)
-import Servant (Context ((:.), EmptyContext))
-import Servant.API (AuthProtect)
-import Servant.Client (BaseUrl, Client, ClientEnv (baseUrl), ClientM, HasClient, ServantError,
-                       runClientM)
-import Servant.Client.Core (RunClient (..), clientIn)
-import Servant.Client.Internal.HttpClient (requestToClientRequest)
-import Servant.Server (Handler, err401, errBody)
-import Servant.Server.Experimental.Auth (AuthHandler, AuthServerData, mkAuthHandler)
-
-import qualified Data.ByteArray as BA (convert)
-import qualified Data.ByteString as BS
-import qualified Data.ByteString.Base64 as Base64
-import qualified Data.ByteString.Lazy as LBS (toStrict)
-import qualified Network.HTTP.Client as Client (host, method, path, queryString, requestBody,
-                                                requestHeaders)
-import qualified Network.Wai as Wai (Request, requestHeaders)
-import qualified Servant.Client.Core as Servant (Request, Response, StreamingResponse,
-                                                 requestHeaders)
-
-----------------------------------------------------------------------------
--- Crypto
-----------------------------------------------------------------------------
-
--- | The wraper for the secret key.
-newtype SecretKey = SecretKey
-    { unSecretKey :: ByteString
-    }
-
--- | Hashed message used as the signature. Encoded in Base64.
-newtype Signature = Signature
-    { unSignature :: ByteString
-    } deriving (Eq)
-
-{- | Compute the hashed message using the supplied hashing function. And then
-encode the result in the Base64 encoding.
--}
-sign :: forall algo . (HashAlgorithm algo)
-     => SecretKey   -- ^ Secret key to use
-     -> ByteString  -- ^ Message to MAC
-     -> Signature   -- ^ Hashed message
-sign (SecretKey sk) msg = Signature
-                        $ Base64.encode
-                        $ BA.convert
-                        $ hmacGetDigest
-                        $ hmac @_ @_ @algo sk msg
-{-# INLINE sign #-}
-
--- | 'sign' function specialized for 'SHA256' cryptographic algorithm.
-signSHA256 :: SecretKey -> ByteString -> Signature
-signSHA256 = sign @SHA256
-{-# INLINE signSHA256 #-}
-
-----------------------------------------------------------------------------
--- Web
-----------------------------------------------------------------------------
-
--- | Part of the HTTP request that will be signed.
-data RequestPayload = RequestPayload
-    { rpMethod  :: !Method  -- ^ HTTP method
-    , rpContent :: !ByteString  -- ^ Raw content of HTTP body
-    , rpHeaders :: !RequestHeaders  -- ^ All headers of HTTP request
-    , rpRawUrl  :: !ByteString  -- ^ Raw request URL with host, path pieces and parameters
-    }
-
-waiRequestToPayload :: Wai.Request -> IO RequestPayload
-waiRequestToPayload req = getWaiRequestBody req >>= \body -> pure RequestPayload
-    { rpMethod  = requestMethod req
-    , rpContent = body
-    , rpHeaders = Wai.requestHeaders req
-    , rpRawUrl  = fromMaybe mempty (requestHeaderHost req) <> rawPathInfo req <> rawQueryString req
-    }
-
-servantRequestToPayload :: BaseUrl -> Servant.Request -> RequestPayload
-servantRequestToPayload url (requestToClientRequest url -> req) = RequestPayload
-    { rpMethod  = Client.method req
-    , rpContent = toBsBody $ Client.requestBody req
-    , rpHeaders = Client.requestHeaders req
-    , rpRawUrl  = Client.host req <> Client.path req <> Client.queryString req
-    }
-  where
-    toBsBody :: RequestBody -> ByteString
-    toBsBody (RequestBodyBS bs)       = bs
-    toBsBody (RequestBodyLBS bs)      = LBS.toStrict bs
-    toBsBody (RequestBodyBuilder _ b) = LBS.toStrict $ toLazyByteString b
-    toBsBody _                        = ""  -- heh
-
--- TODO: require Content-Type header?
--- TODO: require Date header with timestamp?
-{- | This function signs HTTP request according to the following algorithm:
-
-@
-stringToSign = HTTP-Method       ++ "\n"
-            ++ Content-MD5       ++ "\n"
-            ++ HeadersNormalized ++ "\n"
-            ++ RawURL
-
-signature = encodeBase64
-          $ signHmac yourSecretKey
-          $ encodeUtf8 stringToSign
-@
-
-where @HeadersNormalized@ are headers decapitalzed, joined, sorted
-alphabetically and intercalated with line break. So, if you have headers like
-these:
-
-@
-User-Agent: Mozilla/5.0
-Host: foo.bar.com
-@
-
-the result of header normalization will look like this:
-
-@
-hostfoo.bar.com
-user-agentMozilla/5.0
-@
--}
-requestSignature
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key to use
-    -> RequestPayload  -- ^ Payload to sign
-    -> Signature
-requestSignature signer sk = signer sk . createStringToSign
-  where
-    createStringToSign :: RequestPayload -> ByteString
-    createStringToSign RequestPayload{..} = BS.intercalate "\n"
-        [ rpMethod
-        , hashMD5 rpContent
-        , normalizeHeaders rpHeaders
-        , rpRawUrl
-        ]
-
-    normalizeHeaders :: [Header] -> ByteString
-    normalizeHeaders = BS.intercalate "\n" . sort . map normalize
-      where
-        normalize :: Header -> ByteString
-        normalize (name, value) = foldedCase name <> value
-
-
-{- | This function takes signing function @signer@ and secret key and expects
-that given 'Request' has header:
-
-@
-Authentication: HMAC <signature>
-@
-
-It checks whether @<signature>@ is true request signature.
--}
-verifySignatureHmac
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key that was used for signing 'Request'
-    -> RequestPayload
-    -> Bool
-verifySignatureHmac signer sk signedPayload = case unsignedPayload of
-    Nothing         -> False
-    Just (pay, sig) -> sig == requestSignature signer sk pay
-  where
-    -- Extracts HMAC signature from request and returns request with @authHeaderName@ header
-    unsignedPayload :: Maybe (RequestPayload, Signature)
-    unsignedPayload = case extractOn isAuthHeader $ rpHeaders signedPayload of
-        (Nothing, _) -> Nothing
-        (Just (_, val), headers) -> BS.stripPrefix "HMAC " val >>= \sig -> Just
-            ( signedPayload { rpHeaders = headers }
-            , Signature sig
-            )
-
-----------------------------------------------------------------------------
--- Internals
-----------------------------------------------------------------------------
-
-authHeaderName :: HeaderName
-authHeaderName = "Authentication"
-
-isAuthHeader :: Header -> Bool
-isAuthHeader = (== authHeaderName) . fst
-
-hashMD5 :: ByteString -> ByteString
-hashMD5 = BA.convert . hash @_ @MD5
-
-getWaiRequestBody :: Wai.Request -> IO ByteString
-getWaiRequestBody request = BS.concat <$> getChunks
-  where
-    getChunks :: IO [ByteString]
-    getChunks = requestBody request >>= \chunk ->
-        if chunk == BS.empty
-        then pure []
-        else (chunk:) <$> getChunks
-
-{- | Removes and returns first element from list that satisfies given predicate.
-
->>> extractOn (== 3) [1..5]
-(Just 3, [1,2,4,5])
->>> extractOn (== 3) [5..10]
-(Nothing,[5,6,7,8,9,10])
--}
-extractOn :: (a -> Bool) -> [a] -> (Maybe a, [a])
-extractOn p l =
-    let (before, after) = break p l
-    in case uncons after of
-        Nothing      -> (Nothing, l)
-        Just (x, xs) -> (Just x, before ++ xs)
-
-----------------------------------------------------------------------------
--- Servant Server
-----------------------------------------------------------------------------
-
-type HmacAuth = AuthProtect "hmac-auth"
-
-type instance AuthServerData HmacAuth = ()
-
-type HmacAuthContextHandlers = '[AuthHandler Wai.Request ()]
-type HmacAuthContext = Context HmacAuthContextHandlers
-
-hmacAuthServerContext
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key that was used for signing 'Request'
-    -> HmacAuthContext
-hmacAuthServerContext signer sk = hmacAuthHandler signer sk :. EmptyContext
-
-hmacAuthHandler
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key that was used for signing 'Request'
-    -> AuthHandler Wai.Request ()
-hmacAuthHandler signer sk = mkAuthHandler handler
-  where
-    handler :: Wai.Request -> Handler ()
-    handler req = liftIO (verifySignatureHmac signer sk <$> waiRequestToPayload req) >>= \case
-        True  -> pure ()
-        False -> throwError $ err401 { errBody = "HMAC Auth failed." }
-
-----------------------------------------------------------------------------
--- Servant client
-----------------------------------------------------------------------------
-
--- | Environment for 'HmacClientM'.
-data HmacSettings = HmacSettings
-    { hmacSigner    :: SecretKey -> ByteString -> Signature
-    , hmacSecretKey :: SecretKey
-    }
-
-{- | @newtype@ wrapper over 'ClientM' that signs all outgoing requests
-automatically.
--}
-newtype HmacClientM a = HmacClientM
-    { runHmacClientM :: ReaderT HmacSettings ClientM a
-    } deriving (Functor, Applicative, Monad, MonadReader HmacSettings)
-
-hmacifyClient :: ClientM a -> HmacClientM a
-hmacifyClient = HmacClientM . lift
-
-{- | Adds signed header to the request.
-
-@
-Authentication: HMAC <signature>
-@
--}
-signRequestHmac
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key that was used for signing 'Request'
-    -> BaseUrl  -- ^ Base url for servant request
-    -> Servant.Request  -- ^ Original request
-    -> Servant.Request  -- ^ Signed request
-signRequestHmac signer sk url req = do
-    let payload = servantRequestToPayload url req
-    let signature = requestSignature signer sk payload
-    let authHead = (authHeaderName, "HMAC " <> unSignature signature)
-    req { Servant.requestHeaders = authHead <| Servant.requestHeaders req }
-
-hmacClientSign :: Servant.Request -> HmacClientM Servant.Request
-hmacClientSign req = HmacClientM $ do
-    HmacSettings{..} <- ask
-    url <- lift $ asks baseUrl
-    pure $ signRequestHmac hmacSigner hmacSecretKey url req
-
-instance RunClient HmacClientM where
-    runRequest :: Servant.Request -> HmacClientM Servant.Response
-    runRequest = hmacClientSign >=> hmacifyClient . runRequest
-
-    streamingRequest :: Servant.Request -> HmacClientM Servant.StreamingResponse
-    streamingRequest = hmacClientSign >=> hmacifyClient . streamingRequest
-
-    throwServantError :: ServantError -> HmacClientM a
-    throwServantError = hmacifyClient . throwServantError
-
-runHmacClient
-    :: (SecretKey -> ByteString -> Signature)  -- ^ Signing function
-    -> SecretKey  -- ^ Secret key to sign all requests
-    -> ClientEnv
-    -> HmacClientM a
-    -> IO (Either ServantError a)
-runHmacClient hmacSigner hmacSecretKey env client =
-    runClientM (runReaderT (runHmacClientM client) HmacSettings{..}) env
-
--- | Generates a set of client functions for an API.
-hmacClient :: forall api .  HasClient HmacClientM api => Client HmacClientM api
-hmacClient = Proxy @api `clientIn` Proxy @HmacClientM
+import Servant.Auth.Hmac.Client as Hmac
+import Servant.Auth.Hmac.Crypto as Hmac
+import Servant.Auth.Hmac.Server as Hmac
